adjusted structure of App component. moved route to app.jsx and added Switch component
