--- src/App.js
@@ -1,30 +1,0 @@
-import logo from './logo.svg';
-import './App.css';
-import Header from './Header/Header';
-import axios from 'axios';
-
-
-
-
-
-
-function App() {
-
-
-
-
-
-
-  return(
-  <div className="App">
-    
-      <Header />
-
-
-
-
-    </div>
-  );
-}
-
-export default App;
